# openapi/pelican.yaml
info:
  title: pelican
openapi: 3.1.0
paths:
...
  /media/{id}:
    get:
      summary: Get media
      description: Get media by ID.
      operationId: MediaIdGet
      parameters:
        - name: id
          in: path
          schema:
            type: string
            format: uuid
            description: Identifier of the media to get.
          description: Identifier of the media to get.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
        - name: include
          in: query
          schema:
            oneOf:
              - type: "null"
              - type: string
            description: Relations to include in the response.
          description: Relations to include in the response.
          required: false
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
      responses:
        "200":
          description: Request fulfilled, document follows
          headers: {}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/media_models_Media"
        "400":
          description: Bad request syntax or unsupported method
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                  detail:
                    type: string
                  extra:
                    additionalProperties: {}
                    type:
                      - "null"
                      - object
                      - array
                type: object
                required:
                  - detail
                  - status_code
                description: Validation Exception
                examples:
                  - status_code: 400
                    detail: Bad Request
                    extra: {}
      deprecated: false
    delete:
      summary: Delete media
      description: Delete media by ID.
      operationId: MediaIdDelete
      parameters:
        - name: id
          in: path
          schema:
            type: string
            format: uuid
            description: Identifier of the media to delete.
          description: Identifier of the media to delete.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
      responses:
        "204":
          description: Request fulfilled, nothing follows
          headers: {}
          content:
            application/json:
              schema:
                type: "null"
        "400":
          description: Bad request syntax or unsupported method
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                  detail:
                    type: string
                  extra:
                    additionalProperties: {}
                    type:
                      - "null"
                      - object
                      - array
                type: object
                required:
                  - detail
                  - status_code
                description: Validation Exception
                examples:
                  - status_code: 400
                    detail: Bad Request
                    extra: {}
      deprecated: false
    patch:
      summary: Update media
      description: Update media by ID.
      operationId: MediaIdUpdate
      parameters:
        - name: id
          in: path
          schema:
            type: string
            format: uuid
            description: Identifier of the media to update.
          description: Identifier of the media to update.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
        - name: include
          in: query
          schema:
            oneOf:
              - type: "null"
              - type: string
            description: Relations to include in the response.
          description: Relations to include in the response.
          required: false
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUpdateInput"
        required: true
      responses:
        "200":
          description: Request fulfilled, document follows
          headers: {}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/media_models_Media"
        "400":
          description: Bad request syntax or unsupported method
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                  detail:
                    type: string
                  extra:
                    additionalProperties: {}
                    type:
                      - "null"
                      - object
                      - array
                type: object
                required:
                  - detail
                  - status_code
                description: Validation Exception
                examples:
                  - status_code: 400
                    detail: Bad Request
                    extra: {}
      deprecated: false
  /media/{id}/content:
    get:
      summary: Download media content
      description: Download media content by ID.
      operationId: MediaIdContentDownload
      parameters:
        - name: id
          in: path
          schema:
            type: string
            format: uuid
            description: Identifier of the media to download content for.
          description: Identifier of the media to download content for.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
      responses:
        "200":
          description: Stream Response
          headers:
            content-length:
              schema:
                type: string
              description: File size in bytes
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            last-modified:
              schema:
                type: string
                format: date-time
              description: Last modified data-time in RFC 2822 format
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            etag:
              schema:
                type: string
              description: Entity tag
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            Last-Modified:
              description: Last modified.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            Content-Length:
              description: Content length.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            ETag:
              description: Entity tag.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            Content-Type:
              description: Content type.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
          content:
            "":
              schema:
                type: string
                contentMediaType: application/octet-stream
        "400":
          description: Bad request syntax or unsupported method
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                  detail:
                    type: string
                  extra:
                    additionalProperties: {}
                    type:
                      - "null"
                      - object
                      - array
                type: object
                required:
                  - detail
                  - status_code
                description: Validation Exception
                examples:
                  - status_code: 400
                    detail: Bad Request
                    extra: {}
      deprecated: false
    put:
      summary: Upload media content
      description: Upload media content by ID.
      operationId: MediaIdContentUpload
      parameters:
        - name: id
          in: path
          schema:
            type: string
            format: uuid
            description: Identifier of the media to upload content for.
          description: Identifier of the media to upload content for.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
        - name: Content-Type
          in: header
          schema:
            type: string
            description: Content type.
          description: Content type.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
      responses:
        "204":
          description: Request fulfilled, nothing follows
          headers: {}
          content:
            application/json:
              schema:
                type: "null"
        "400":
          description: Bad request syntax or unsupported method
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                  detail:
                    type: string
                  extra:
                    additionalProperties: {}
                    type:
                      - "null"
                      - object
                      - array
                type: object
                required:
                  - detail
                  - status_code
                description: Validation Exception
                examples:
                  - status_code: 400
                    detail: Bad Request
                    extra: {}
      deprecated: false
    head:
      summary: Get media content headers
      description: Get media content headers by ID.
      operationId: MediaIdContentHeaddownload
      parameters:
        - name: id
          in: path
          schema:
            type: string
            format: uuid
            description: Identifier of the media to get content headers for.
          description: Identifier of the media to get content headers for.
          required: true
          deprecated: false
          allowEmptyValue: false
          allowReserved: false
      responses:
        "200":
          description: Request fulfilled, document follows
          headers:
            Last-Modified:
              description: Last modified.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            Content-Length:
              description: Content length.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            ETag:
              description: Entity tag.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
            Content-Type:
              description: Content type.
              required: false
              deprecated: false
              allowEmptyValue: false
              allowReserved: false
          content:
            application/json:
              schema:
                type: "null"
        "400":
          description: Bad request syntax or unsupported method
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                  detail:
                    type: string
                  extra:
                    additionalProperties: {}
                    type:
                      - "null"
                      - object
                      - array
                type: object
                required:
                  - detail
                  - status_code
                description: Validation Exception
                examples:
                  - status_code: 400
                    detail: Bad Request
                    extra: {}
      deprecated: false
...
components:
  schemas:
...
    MediaUpdateInput:
      properties:
        id:
          type: string
        name:
          type: string
      type: object
      required: []
      title: MediaUpdateInput
      description: Data to update media.
...
    media_models_Binding:
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the binding.
        playlistId:
          type: string
          format: uuid
          description: Identifier of the playlist that the binding belongs to.
        mediaId:
          type: string
          format: uuid
          description: Identifier of the media that the binding belongs to.
        rank:
          type: string
          description: Rank of the media in the binding.
        playlist:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/media_models_Playlist"
          description: Playlist that the binding belongs to.
        media:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/media_models_Media"
          description: Media that the binding belongs to.
      type: object
      required:
        - id
        - mediaId
        - playlistId
        - rank
      title: Binding
    media_models_Media:
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the media.
        name:
          type: string
          description: Name of the media.
        bindings:
          oneOf:
            - type: "null"
            - items:
                $ref: "#/components/schemas/media_models_Binding"
              type: array
          description: Bindings that the media belongs to.
      type: object
      required:
        - id
        - name
      title: Media
    media_models_Playlist:
      properties:
        id:
          type: string
          format: uuid
          description: Identifier of the playlist.
        name:
          type: string
          description: Name of the playlist.
        bindings:
          oneOf:
            - type: "null"
            - items:
                $ref: "#/components/schemas/media_models_Binding"
              type: array
          description: Bindings that the playlist belongs to.
      type: object
      required:
        - id
        - name
      title: Playlist
...